% Generated by roxygen2 (4.0.1): do not edit by hand
\name{odbcFetchRows}
\alias{odbcFetchRows}
\title{Overlay over \link[RODBC]{odbcFetchRows}}
\usage{
odbcFetchRows(channel, ...)
}
\arguments{
\item{channel}{ODBC connection obtained by \link{odbcConnect}}

\item{...}{other parametrs passed to \link[RODBC]{odbcFetchRows}}
}
\value{
see \link[RODBC]{odbcFetchRows}
}
\description{
RODBC::odbcFetchRows crashes if the ODBC channel is in "query prepared
but already not executed" state.
This function is a small overlay emmitting an error in such a case.
}

