% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odbcQueryTimeout.R
\name{odbcGetQueryTimeout}
\alias{odbcGetQueryTimeout}
\title{Get the current query timeout of a prepared query}
\usage{
odbcGetQueryTimeout(channel)
}
\arguments{
\item{channel}{an RODBC channel containing an open connection}
}
\value{
The current query timeout value in seconds. 0 means "no timeout"
}
\description{
Throws an error if an error occured
}
\details{
A query has to be already prepared using SQLPrepare()
}
\seealso{
\code{\link{odbcSetQueryTimeout}}, \code{\link{odbcConnect}}, \code{\link{odbcDriverConnect}}
}

