% Generated by roxygen2 (4.0.1): do not edit by hand
\name{sqlExecute}
\alias{sqlExecute}
\title{Executes an already prepared query}
\usage{
sqlExecute(channel, data = NULL, fetch = FALSE, errors = TRUE,
  rows_at_time = attr(channel, "rows_at_time"), ...)
}
\arguments{
\item{channel}{ODBC connection obtained by odbcConnect()}

\item{data}{data to pass to sqlExecute (as data.frame)}

\item{fetch}{whether to automatically fetch results (if data provided)}

\item{errors}{whether to display errors}

\item{rows_at_time}{number of rows to fetch at one time - see details of \link{sqlQuery}}

\item{...}{parameters to pass to \link{sqlFetchMore} (if fetch=TRUE)}
}
\value{
see datails
}
\description{
Executes an already prepared query. You may prepare query using \link{sqlPrepare}.

Optionally fetches results using \link{sqlFetch}.
}
\details{
Return value depends on the combination of parameters:
\itemize{
  \item if there were errors during query execution (or fetching results)
    return value depends on errors parameter - if errors=TRUE, the vector of error messages
    will be returned, otherwise -1 will be returned
  \item if fetch=FALSE and there were no errors during query execution, invisible(1) will be returned
  \item if fetch=TRUE and there were no errors during query execution and fetching results,
    data.frame with results will be returned
}
}
\examples{
\dontrun{
  conn = odbcConnect('MyDataSource')

  # prepare, execute and fetch results separatly
  sqlPrepare(conn, "SELECT * FROM myTable WHERE column = ?")
  sqlExecute(conn, 'myValue')
  sqlFetch(conn)

  # prepare and execute at one time, fetch results separately
  sqlPrepare(conn, "SELECT * FROM myTable WHERE column = ?", 'myValue')
  sqlFetchMore(conn)

  # prepare, execute and fetch at one time
  sqlPrepare(conn, "SELECT * FROM myTable WHERE column = ?", 'myValue', TRUE)

  # prepare, execute and fetch at one time, pass additional parameters to sqlFetch()
  sqlPrepare(conn, "SELECT * FROM myTable WHERE column = ?", 'myValue', TRUE, stringsAsFactors=FALSE)
}
}

