% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlFetchMore.R
\name{sqlFetchMore}
\alias{sqlFetchMore}
\title{Overlay over \link[RODBC]{sqlFetchMore}}
\usage{
sqlFetchMore(channel, ...)
}
\arguments{
\item{channel}{ODBC connection obtained by \link{odbcConnect}}

\item{...}{other parametrs passed to \link[RODBC]{sqlFetchMore}}
}
\value{
see \link[RODBC]{sqlFetchMore}
}
\description{
RODBC::sqlFetchMore crashes if the ODBC channel is in "query prepared
but already not executed" state.
This function is a small overlay emmitting an error in such a case.
}
