% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odbcQueryTimeout.R
\name{odbcSetQueryTimeout}
\alias{odbcSetQueryTimeout}
\title{Sets the query timeout of a prepared query}
\usage{
odbcSetQueryTimeout(channel, timeout = 30)
}
\arguments{
\item{channel}{an open RODBC channel (connection)}

\item{timeout}{the new query timeout value in seconds (0 means "no timeout")}
}
\value{
0 = success,
         1 = success but with an info message,
}
\description{
Throws an error if an error occured
}
\details{
A query has to be already prepared using SQLPrepare()
}
\seealso{
\code{\link{odbcSetQueryTimeout}}, \code{\link{odbcConnect}}, \code{\link{odbcDriverConnect}}
}

